refactor(cli): split template into api/template leaf shape
Pure reorg — `--json` and human output stay byte-identical. Move shared
discovery/IO into api/template/_adapter.mjs, split the command modes into
list/show/skeleton/copy leaves, and make template.mjs a dispatcher + barrel
that re-exports every previously-exported symbol (template, discoverTemplates,
discoverAll, discoverAllWithErrors, discoverIntegrationTemplatesForOne,
findShowcase, findRelatedBlocks, stripTemplateAssetRefs, listTemplates,
extractComponents, plus the DiscoveredTemplate/TemplateDiscoveryError types).

template is a shared module: component/layout/search/init/discover/
validate-integration and lib/project import helpers from
api/template/template.mjs, so preserving the barrel's exports keeps them
resolving unchanged. Leaves carry precise @returns; types stay central.
